프로그래머스 레벨2 후보키 문제 풀이 완료.
많이 헤맴, 조합 관련 공식도 한 번쯤 확인해보는게 좋을듯

점점 알고리즘 푸는게 스트레스가 아니라 재미가 있음 ^^

+++ src/Programmers/Level2_KAKAO_CACHE.java
@@ -56,7 +56,7 @@ public class Level2_KAKAO_CACHE {
 }
 
 class CacheNode implements Comparable<CacheNode> {
-    private int hitNum;
+    private long hitNum;
     private String city;
 
     public CacheNode(int hitNum, String city) {
